+++ src/pages/GroupEdit.vue
@@ -2,6 +2,12 @@
   <v-container>
     <ProgressCircular v-if="status === 'loading'" />
     <LoadFailedText v-else-if="status === 'error'" />
+    <template v-else-if="!canEdit">
+      <v-icon large color="error" class="mr-5">mdi-alert-circle</v-icon>
+      <span class="text--secondary headline">
+        traP公式のグループは編集できません
+      </span>
+    </template>
     <template v-else>
       <v-stepper v-model="step" class="mb-5">
         <v-stepper-header>
@@ -34,7 +40,14 @@
         </v-stepper-items>
       </v-stepper>
 
-      <DeleteConfirmationDialog v-model="dialog" @confirm="deleteGroup" />
+      <v-card class="px-5 pt-5 pb-3">
+        <span class="headline mr-3">
+          Delete this group
+        </span>
+        <v-btn small depressed color="error" class="mb-2" @click="deleteGroup">
+          Delete
+        </v-btn>
+      </v-card>
     </template>
   </v-container>
 </template>
@@ -49,7 +62,6 @@ import FormBackButton from '@/components/shared/FormBackButton.vue'
 import ProgressCircular from '@/components/shared/ProgressCircular.vue'
 import LoadFailedText from '@/components/shared/LoadFailedText.vue'
 import RepositoryFactory from '@/repositories/RepositoryFactory'
-import DeleteConfirmationDialog from '@/components/shared/DeleteConfirmationDialog.vue'
 
 const GroupsRepo = RepositoryFactory.get('groups')
 
@@ -61,29 +73,34 @@ const GroupsRepo = RepositoryFactory.get('groups')
     FormBackButton,
     ProgressCircular,
     LoadFailedText,
-    DeleteConfirmationDialog,
   },
 })
 export default class GroupEdit extends Vue {
   status: 'loading' | 'loaded' | 'error' = 'loading'
-  dialog = false
+  canEdit = true
   valid = false
   step = 1
 
   group: Schemas.Group | null = null
 
-  created() {
-    this.fetchGroupData()
-  }
-
-  async fetchGroupData() {
+  async created() {
     this.status = 'loading'
     const groupId = this.$route.params.id
     try {
-      this.group = (await GroupsRepo.$groupId(groupId).get()).data
+      this.fetchGroupData()
       this.status = 'loaded'
     } catch (__) {
       this.status = 'error'
+    }
+  }
+
+  async fetchGroupData() {
+    const groupId = this.$route.params.id
+    const group = (await GroupsRepo.$groupId(groupId).get()).data
+    if (group.isTraQGroup) {
+      this.canEdit = false
+    } else {
+      this.group = group
     }
   }
 
@@ -99,6 +116,10 @@ export default class GroupEdit extends Vue {
   }
 
   async deleteGroup() {
+    const confirmed = window.confirm(
+      'この操作は取り消せません。本当にこのグループを削除してもよろしいですか？'
+    )
+    if (!confirmed) return
     const groupId = this.$route.params.id
     try {
       await GroupsRepo.$groupId(groupId).delete()
